requirements.txt
vFlask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
pandas==2.1.0
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
python-dotenv==1.0.0
gunicorn==21.2.0
plotly==5.16.1
dash==2.11.1
dash-bootstrap-components==1.4.2 